fix(dashboard & data-discovery): fix type of time related fields from `long` to `OffsetDateTime` or `Double`

+++ kun-data-discovery/src/main/java/com/miotech/kun/datadiscovery/model/entity/Watermark.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class Watermark {
-
-    Long time;
+    /* Epoch Unix timestamp in milliseconds */
+    Double time;
 }
